feat: add custom confirm-dialog component

+++ src/app/(core)/workouts/workout-active/delete-active-workout-dialog.tsx
@@ -1,15 +1,8 @@
 "use client";
 
 import { useDeleteActiveWorkout } from "@/api/workouts/mutations";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
+import ConfirmDialog from "@/components/ui/confirm-dialog";
+
 import { useActiveWorkoutContext } from "@/context/active-workout-context";
 
 export default function DeleteActiveWorkoutDialog({
@@ -23,27 +16,16 @@ export default function DeleteActiveWorkoutDialog({
   const { setActiveWorkoutOpen } = useActiveWorkoutContext();
 
   return (
-    <Dialog open={isOpen} onOpenChange={onOpenChanged}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Discard workout?</DialogTitle>
-        </DialogHeader>
-        <DialogFooter className="grid grid-cols-2">
-          <DialogClose asChild>
-            <Button variant="outline">Close</Button>
-          </DialogClose>
-          <Button
-            variant="destructive"
-            onClick={() => {
-              setActiveWorkoutOpen(false);
-              deleteActiveWorkoutMutation();
-              onOpenChanged(false);
-            }}
-          >
-            Delete
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
+    <ConfirmDialog
+      isOpen={isOpen}
+      onOpenChange={onOpenChanged}
+      onConfirm={() => {
+        setActiveWorkoutOpen(false);
+        deleteActiveWorkoutMutation();
+        onOpenChanged(false);
+      }}
+      title="Discard Workout"
+      variant="destructive"
+    />
   );
 }
